Fix font glyphs

+++ main/global_state.h
@@ -33,9 +33,9 @@ typedef struct
     int historical_hashrate_init;
     double current_hashrate;
     int64_t start_time;
-    uint64_t shares_accepted;
-    uint64_t shares_rejected;
-    uint64_t work_received;
+    uint32_t shares_accepted;
+    uint32_t shares_rejected;
+    uint32_t work_received;
     RejectedReasonStat rejected_reason_stats[10];
     int rejected_reason_stats_count;
     int screen_page;
